Restore repository to 66b217b (no history rewrite) [skip ci]
