+++ build.gradle
@@ -48,7 +48,7 @@ dependencies {
 	// dummy data 용
 	implementation 'net.datafaker:datafaker:2.4.4'
 
-	//..env 인식
+	//.env 인식
 	implementation 'io.github.cdimascio:java-dotenv:5.2.2'
 
 	// FCM
