+++ docker-compose-dev.yml
@@ -12,7 +12,7 @@ services:
       context: .
       dockerfile: Dockerfile
     env_file:
-      - ..env
+      - .env
     restart: always
     ports:
       - "8000:8000"
